services/rag-service/requirements.txt
# RAG Service Dependencies
# This service now includes the full rag_system module
# Updated: 2025-12-08 - Versions matched to latest

# FastAPI Framework
fastapi==0.124.0
uvicorn==0.38.0
starlette==0.50.0

# Pydantic for validation
pydantic==2.12.5
pydantic-settings==2.12.0

# Core LangChain and RAG Dependencies
langchain==1.1.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langchain-experimental==0.4.0

# LLM Integration - Ollama
langchain-ollama==1.0.0

# Vector Database
chromadb==1.3.5
python-multipart==0.0.20

# Document Processing
python-docx==1.2.0
python-magic==0.4.27
pillow==11.3.0

# Docling for Advanced Document Processing
docling==2.64.0
langchain-docling==2.0.0
docling-core==2.54.0

# Configuration
PyYAML==6.0.3

# Hybrid Search (BM25 + Vector)
rank-bm25==0.2.2

# Pipeline Orchestration
prefect==3.6.5
py7zr>=0.22.0

# HTTP Client
httpx==0.28.1

# Utilities
tqdm==4.67.1
rich==14.2.0
